refactor: mark selection
